+++ examples/torch_code/train_fully_connected.py
@@ -40,10 +40,10 @@ def sigma(x):
 class FullyConnectedNetwork(nn.Module):
     def __init__(self, W, B, w, b, n):
         super().__init__()
-        self.W = nn.Parameter(torch.as_tensor(W).float())
-        self.B = nn.Parameter(torch.as_tensor(B).float())
-        self.w = nn.Parameter(torch.as_tensor(w).float())
-        self.b = nn.Parameter(torch.as_tensor(b).float())
+        self.W = nn.Parameter(torch.as_tensor(W))
+        self.B = nn.Parameter(torch.as_tensor(B))
+        self.w = nn.Parameter(torch.as_tensor(w))
+        self.b = nn.Parameter(torch.as_tensor(b))
         self.n = torch.as_tensor(n).float() if isinstance(n, (int, float, torch.Tensor)) else n
 
     def forward(self, x):
